+++ aider/coders/__init__.py
@@ -1,11 +1,11 @@
+from .architect_coder import ArchitectCoder
 from .ask_coder import AskCoder
 from .base_coder import Coder
 from .editblock_coder import EditBlockCoder
 from .editblock_fenced_coder import EditBlockFencedCoder
+from .editor_editblock_coder import EditorEditBlockCoder
+from .editor_whole_coder import EditorWholeFileCoder
 from .help_coder import HelpCoder
-from .junior_editblock_coder import JuniorEditBlockCoder
-from .junior_whole_coder import JuniorWholeFileCoder
-from .senior_coder import SeniorCoder
 from .udiff_coder import UnifiedDiffCoder
 from .wholefile_coder import WholeFileCoder
 
@@ -20,7 +20,7 @@ __all__ = [
     WholeFileCoder,
     UnifiedDiffCoder,
     #    SingleWholeFileFunctionCoder,
-    SeniorCoder,
-    JuniorEditBlockCoder,
-    JuniorWholeFileCoder,
+    ArchitectCoder,
+    EditorEditBlockCoder,
+    EditorWholeFileCoder,
 ]