+++ aider/main.py
@@ -536,8 +536,8 @@ def main(argv=None, input=None, output=None, force_git_root=None, return_coder=F
     main_model = models.Model(
         args.model,
         weak_model=args.weak_model,
-        junior_model=args.junior_model,
-        junior_edit_format=args.junior_edit_format,
+        editor_model=args.editor_model,
+        editor_edit_format=args.editor_edit_format,
     )
 
     if args.verbose: